Petya: new tables/models(Company, Position) are ready, dummy users have company and position
