+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
 spring.datasource.url=jdbc:postgresql://localhost/bernard_db
-spring.datasource.username=dry
-spring.datasource.password=
+spring.datasource.username=petya
+spring.datasource.password=petya
 spring.jpa.generate-ddl=true
